Move `HashMethod` to separate file
